feat(l2): add a blobs saver command to store state diffs (#2194)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We need a tool to store state diffs blobs offline so the L2 state is
reconstructable after 2 weeks, when blobs got deleted on L1.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
This PR adds a command to the L2 CLI (`ethrex_l2 stack blobs-saver`)
that runs a service which continuously looks for new Commit events in
the `OnChainProposer` contract and downloads its blobs in the local
filesystem.
It uses a both EL and CL RPCs

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1196

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>