feat(l2): pico prover (#1922)
**Motivation**

Adds [Pico](https://pico-docs.brevis.network/) as a prover backend. 

Also does a major refactor to remove the need to have multiple zkvm
dependencies compiling at the same time, this is because Pico doesn't
compile while also having Risc0 as dependency; the linker fails with a
"duplicated symbols" error.

Also removes zkvm dependencies from crates that don't need them by
decoupling return types. This is because Pico compiles with nightly only
and we want to minimize the number of crates that depend on it (now only
`ethrex-prover` and `zkvm_interface` does)

**Description**

- adds pico as prover backend
- decouples zkvm dependencies from other L2 crates by doing a major
refactor of provers
- makes it so you can compile the prover client with only one backend at
a time
- makes the prover client return the proof calldata to send to the L1
bridge contract instead of the proofs using each custom type of every
zkvm

---------

Co-authored-by: Mario Rugiero <mrugiero@gmail.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>