refactor(core): ethrex cli (#2240)
**Motivation**

To improve `ethrex`'s CLI readability and extensibility.

**Description**

This PR refactors de CLI to use clap derive instead of clap builder
approach. Using the latter suited perfectly for the first version but as
we keep adding flags/args and subcommands, using the the first is better
for readability and also extensibility.

In the new design, the CLI is modeled as the struct `CLI` as follows:

```Rust
pub struct CLI {
    #[clap(flatten)]
    pub opts: Options,
    #[cfg(feature = "based")]
    #[clap(flatten)]
    pub based_opts: BasedOptions,
    #[command(subcommand)]
    pub command: Option<Subcommand>,
}
```

where `opts` are the flags corresponding to `ethrex` common usage,
`based_opts` are the flags needed when running `ethrex` with the `based`
feature, and `command` is an enum containing the subcommands
(`removedb`, and `import` for now) which is optional.

If you'd want to add a new subcommand, simply add it to the `Subcommand`
enum and implement its handler in the `Subcommand::run` `match`.

The CLI args are contained in `Options` and `BasedOptions`. Adding a new
flag/arg would mean to add a field on the corresponding struct, and if
you want for example to add flags/args for the L2 feature it'd be good
for you to create an `L2Options` struct with them. The
`#[clap(flatten)]` basically "unpacks" the struct fields (args and
flags) for the CLI.

Running `cargo run --release --bin ethrex -- --help` displays:

```Shell
Usage: ethrex [OPTIONS] [COMMAND]

Commands:
  removedb  Remove the database
  import    Import blocks to the database
  help      Print this message or the help of the given subcommand(s)

Options:
  -h, --help
          Print help (see a summary with '-h')

  -V, --version
          Print version

RPC options:
      --http.addr <ADDRESS>
          Listening address for the http rpc server.

          [default: localhost]

      --http.port <PORT>
          Listening port for the http rpc server.

          [default: 8545]

      --authrpc.addr <ADDRESS>
          Listening address for the authenticated rpc server.

          [default: localhost]

      --authrpc.port <PORT>
          Listening port for the authenticated rpc server.

          [default: 8551]

      --authrpc.jwtsecret <JWTSECRET_PATH>
          Receives the jwt secret used for authenticated rpc requests.

          [default: jwt.hex]

Node options:
      --log.level <LOG_LEVEL>
          Possible values: info, debug, trace, warn, error

          [default: INFO]

      --network <GENESIS_FILE_PATH>
          Alternatively, the name of a known network can be provided instead to use its preset genesis file and include its preset bootnodes. The networks currently supported include Holesky, Sepolia and Mekong.

      --datadir <DATABASE_DIRECTORY>
          If the datadir is the word `memory`, ethrex will use the `InMemory Engine`.

          [default: ethrex]

      --metrics.port <PROMETHEUS_METRICS_PORT>


      --dev
          If set it will be considered as `true`. The Binary has to be built with the `dev` feature enabled.

      --evm <EVM_BACKEND>
          Has to be `levm` or `revm`

          [default: revm]

P2P options:
      --p2p.enabled


      --p2p.addr <ADDRESS>
          [default: 0.0.0.0]

      --p2p.port <PORT>
          [default: 30303]

      --discovery.addr <ADDRESS>
          UDP address for P2P discovery.

          [default: 0.0.0.0]

      --discovery.port <PORT>
          UDP port for P2P discovery.

          [default: 30303]

      --bootnodes <BOOTNODE_LIST>...
          Comma separated enode URLs for P2P discovery bootstrap.

      --syncmode <SYNC_MODE>
          Can be either "full" or "snap" with "full" as default value.

          [default: full]
```