refactor(l2): add flag for setting sponsor private key (#2281)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

The current implementation requires a `.env` file to exist and the
execution panics if this file does not exist. Nevertheless, this has a
purpose of being. As this feature should be used in `l2` it is assumed
that there's a `.env` file and that is ok because it should. This PR
intends to add a second path for setting the sponsor pk without needing
a `.env`.

**Description**

Add a flag `--sponsor-private-key` as a second option for setting this
value.