chore(core): add `rust-toolchain.toml` (#2278)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We have a pinned version of Rust in the CI and also in `.tool-versions`
(for `asdf`) but not for `rustup`. We encountered ourselves running
different versions of Rust, with different results, specially when
running tools like Clippy

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Added a `rust-toolchain.toml` file with the pinned version of Rust so
it's evaluated by default when using `rustup`. As a side effect, needed
to change the way Pico CLI is installed in the CI.