feat(l1): write multiple account's storage batches in the same db txn (#2270)
**Motivation**
When measuring time taken by each task during snap sync I noticed that a
lot of time was spent writing the storage ranges obtained from peers to
the DB snapshot. It would take anywhere from 3 to over 10 seconds to
write all the ranges to the DB (around 300 storage ranges per request).
This PR modifies the insertion logic to write all 300 ranges in the same
DB transaction, reducing the time taken to write all the ranges to the
DB to 10 milliseconds or less
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add `write_storage_snapshot_batches` method to `Store`, which can
write multiple batches from different accounts on the same txn
* Write all storage ranges received from peers in a single DB txn using
the method above on the storage fetcher (snap sync)
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes: None, but helps speed up snap sync