fix(l1, l2): fix load tests (#2323)
**Motivation**

Load tests were broken for two reasons:

- We were not correctly passing the nonce as an override and thus were
relying on the RPC endpoint to get it, which was not correct (since we
want to pre-send transactions with higher nonces)
- We were hardcoding gas fees; this is because when we first wrote the
load tests, the gas price endpoint on ethrex did not work properly. Now
that it does, we can remove the hardcoded values and just rely on the
endpoint (the default behaviour if you do not pass an `Override` to the
`build_eip1559_transaction` function).

I also changed the `debug!` log when a mempool transaction failed to be
executed while building a block to be an `error!`, because I noticed
it's quite a common occurrence when we run load tests due to some nonce
issue, and I think it's worth investigating (it's the reason why
sometimes we get empty blocks when running load tests).

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: JereSalo <jeresalo17@gmail.com>