fix(core): make metrics port flag not optional (#2343)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Currently, the metrics are initiated iif the `--metrics-port` flag is
passed. This is wrong because the flag is used both to configure the
listening port and as a metrics enabler flag.

If needed, a flag `--metrics.enabled` could be introduced in another PR
if metrics are unwanted for some reason. IMHO initializing metrics as
default is ok.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

- Always initialize metrics
- The `--metrics.port` flag is now not optional and defaults to 9090 as
the default value.