build(l2): enable exec prover by default. (#2372)
**Motivation**
Running `cargo build --workspace` should work by default, without having
to explicitly set a feature flag. Without this change, it errors because
no prover backend was selected. Unless we're working on the prover, we
don't really care about the backend and we should reduce the friction to
people working in the project.

---------

Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>