fix(l1): resume sync upon node restart (#2415)
**Motivation**
After PR #2303 sync cycles are restarted automatically by the same
process that started the sync, removing the notion of "pending syncs".
This works fine until the node is restarted during the sync.
If the node is restarted, no one is checking whether there was an active
sync progress before, and the node continues unknowingly operating on
the invalid state left from the unfinished sync.
This PR aims to fix this by resuming the sync process as soon as the
sync manager is created
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* When creating the SyncManager, check if there are checkpoints from an
active sync process leftover and start the next sync cycle if needed
* in `SyncManager::start_sync` check that the latest fcu head is not the
default before starting a cycle and wait around fro the next fcu update
if needed
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Mario Rugiero <mrugiero@gmail.com>