feat(l1): abstract common fetcher logic used in snap sync (#2408)
**Motivation**
During snap sync, we download account ranges and then for each
downloaded account we request its storage and bytecodes. For these
requests we use fetcher processes that receive incoming messages from a
channel (storage roots, bytecode hashes, etc), place them on a queue,
and then group them in batches and spawn parallel processes to fetch
them. All fetchers share a common behaviour of reading requests,
batching, and fetching with differences concerning only the content of
the queue. In many cases, we have had many bugs due to how these
fetchers worked as we may update one of them and forget about the rest.
This PR aims to reduce the sources of bugs and maintain a unified
behaviour for fetchers by adding generic functions that represent the
fetcher behaviour.
In this PR we add the generic function `run_queue` that receives a
generic queue (a Vec<T>), and an asyn function that operates over a
batch in said queue.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add generic function `run_queue` to abstract queue logic from fetcher
processes
* Use `run_queue` in `bytecode_fetcher`, `large_storage_fecther`, and
`storage_fetcher`

*Considerations*
* As this PR was done with #2359 this won't be applied to the
`storage_healer` which will stop reading messages
* While the batch size could be a const generic instead of a regular
argument, doing so would force us to make the other generic arguments in
`run_queue` explicit, which looks pretty bad
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Mario Rugiero <mrugiero@gmail.com>