fix(levm): try fix blockchain tests (#2436)
**Motivation**
- Fix most blockchain EF Tests with LEVM.

<!-- Why does this pull request exist? What are its goals? -->

**Description**

Every change made here fixes some tests:
- our gas_used is now like REVM's, no more subtracting gas_used -
refunded outside the vm to get gas used.
- there was a little issue with the system_address when we made a
generic contract call, we were deleting it from cache when in fact we
should only delete it if it didn't exist before.
- I also implemented a backup for the coinbase account to restore it's
state when making a generic contract call so that it doesn't change.
- We had differences in `get_state_transitions` with REVM that did not
impact the state but reducing these differences solved some tests. It
was simply that we were saying that an EOF had code: None, while with
REVM we say code: Some(b“”). The error was kind of silly and I could
have fixed it outside the `get_state_transitions` but that was the quick
fix. However, it is always good to return the same as the other
implementation because it enables us to debug better!
- I fixed gas consumption issue in Prague transactions, we had it wrong
in LEVM. I find it strange that the STF EFTests do not test this. I
refactored the whole gas consumption in LEVM's `finalize_execution()`
because it was kinda messy.
- Fix account removal in LEVM: we were removing the receiver account
from the cache when reverting a transaction. This made sense when we
used the cache just for executing one transaction because we didn't want
to modify the receiver, but in any other scenario it was a mistake!
- Fix blob base fee calculation in opcode `blobbasefee`. We were
calculating it in Pectra with Cancun values.

Result: all tests pass except one prague 7702 test and one stack
overflow test.