fix(levm): fix last blockchain ef test (#2455)
**Motivation**

Here we fix the last ef test from the blockchain test suite.

**Description**

When calling a precompiled contract we were returning an execution
report with gas refunded equal to zero. Before Pectra this wasn't a
problem because there wasn't a case were a refund could be made.
But with de EIP-7702, in the `prepare_execution` hook a refund was
possible so that behavior had to be changed.

This branch follows the PR #2436 fixing the test mentioned there.

Closes #2449

---------

Co-authored-by: JereSalo <jeresalo17@gmail.com>