ci(l1): enable blockchain ef tests to be run with levm (#2440)
**Motivation**

We want to run the Ethereum Foundation blockchain tests with LEVM.
Currently, we are only running them with REVM.

**Description**

This PR modifies the EF tests runner so it executes the EF tests with
both VMs. The implementation combines both executions on the same
command `cargo test`, but could be easily modified to include a feature
flag to separate both executions if that's desired.