fix(l1): don't cancel storage healer if state healing ends earlier (#2464)
**Motivation**
Currenlty, when state healing finishes, the storage healer is cancelled,
even if there are still storages to be fetched and the tries are not
stale yet, which slows down storage healing due to the frequent restarts
if state healing ends before storage healing. This PR aims to fix this
by changing the behaviour to not cancel storage healing if state healing
is complete, but to instead give this information to the storage healer
via an AtomicBoo, so that it can decide whether to stop based on if the
state has become stale, or if no more paths are left to heal and mo more
will be added by the state healer.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add AtomicBool param to `storage_healer` to signal whether state
healing has ended
* Don't cancel storage healing if state healer ended due to state
healing being complete
* End storage healer if there are no more pending storages left & state
healing has ended
* (Misc) Remove outdated comment
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number