fix(levm): improve get state transitions LEVM (#2518)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Make `get_state_transitions` we use un LEVM return the same as the one
in REVM.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- I made changes in the past to make `get_state_transitions` for both
LEVM and REVM the same for comparison but I missed one aspect. We only
want to show the code in an `AccountUpdate` if the code itself has been
modified, not just the `AccountInfo`. Before we were returning the code
in the `AccountUpdate` even if only the nonce of the contract changed
for example.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number