feat(levm): improve state EF Tests (#2490)
**Motivation**

- Make running and modifying EF Tests a better experience

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- Remove spinner and just use prints
- We now can filter the tests we want to run by fork name(we do the
filtering in the parsing). Default is all forks.
- Upgrade tests to more recent version
- Run some Legacy Tests that we weren't running before. This adds a lot
of tests more, it is the folder Cancun under LegacyTests. There will be
repeated tests with the folder GeneralStateTests, we may want to find a
solution for that so that it takes less time to execute.
- Create docs in `README.md`
- Implement some nits in the runner, making code easier to understand.
- Ignore a few tests that take too long to run so that we can check for
breaking changes fast.
- Fix comparison report against LEVM, they weren't working correctly
mostly because we were mishandling our Cache
- Tidy the report, now it is much more clear and easier for debugging.
Also the code is easier to follow and more concise too!
- Fix some tests with REVM, basically now using constructor of
`BlobExcessGasAndPrice` and setting chain id to 1 (as we do in LEVM).
- Changed `get_state_transitions`
[here](https://github.com/lambdaclass/ethrex/pull/2518) so that REVM and
LEVM account updates are mostly the same and the comparison is more
accurate for the person who is debugging any test.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number